+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## trajr 1.4.0.9000
 
+* Fix documentation of `TrajDirectionalChange` and `TrajAngles`. The documentation incorrectly stated that the number of angles returned was number of points - 1. It now states that the number of angles is number of points - 2 (which is number of segments - 1).
+
 ## trajr 1.4.0
 
 * Allow the creation of trajectories with 0 points in `TrajFromCoords`.